backend/requirements.txt
asgiref==3.11.1
Django==5.2.2
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
sqlparse==0.5.5

djangorestframework==3.15.1
djangorestframework-simplejwt==5.4.0
Pillow==11.3.0